Fix error '_schema: invalid input' in Django v3.1
This seems to come down to `JSONField` moving to `django.db.models`
(except for back-compat for migrations, and deprecated through to v4)
and `to_python` no longer parsing the JSON, just returning a string.

Unclear to me (having only had a cursory look) why that was removed,
but it was there to begin with, so I assume it's deliberate; so also
switching to `from_db_value` here.